Merge dev to main

+++ src/chocolate_app/routes/users.py
@@ -4,29 +4,28 @@ import base64
 import io
 
 from PIL import Image
-from flask import Blueprint, jsonify, request, abort
+from flask import Blueprint, Response, jsonify, request, abort, send_file
 from werkzeug.security import generate_password_hash
 
 from chocolate_app import DB, get_dir_path, all_auth_tokens, IMAGES_PATH
 from chocolate_app.tables import Users, InviteCodes
-from ..utils.utils import check_authorization, generate_log
-
+from chocolate_app.utils.utils import check_authorization, generate_log
+from chocolate_app.plugins_loader.events import Events, execute_event
 
 dir_path = get_dir_path()
 users_bp = Blueprint("users", __name__)
-
 
 @users_bp.route("/get_all_users", methods=["GET"])
 def get_all_users():
     all_users = Users.query.filter().all()
     all_users_list = []
     for user in all_users:
-        profil_picture = user.profil_picture
-        if not os.path.exists(dir_path + profil_picture):
-            profil_picture = "/static/img/avatars/defaultUserProfilePic.png"
+        profile_picture = user.profile_picture
+        if not os.path.exists(dir_path + profile_picture):
+            profile_picture = "/static/img/avatars/defaultUserProfilePic.png"
         user_dict = {
             "name": user.name,
-            "profil_picture": profil_picture,
+            "profile_picture": profile_picture,
             "account_type": user.account_type,
             "password_empty": True if not user.password else False,
             "id": user.id,
@@ -34,34 +33,62 @@ def get_all_users():
         all_users_list.append(user_dict)
     return jsonify(all_users_list)
 
+@users_bp.route("/profile_picture/<id>", methods=["GET"])
+def get_profile_picture(id):
+    user = Users.query.filter_by(id=id).first()
+    if not user:
+        return send_file("static/img/avatars/defaultUserProfilePic.png")
+    
+    profile_picture = user.profile_picture
+    if not os.path.exists(profile_picture):
+        profile_picture = "static/img/avatars/defaultUserProfilePic.png"
+    return send_file(profile_picture)
 
 @users_bp.route("/login", methods=["POST"])
 def login():
     from uuid import uuid4
 
     auth_token = str(uuid4())
-    account_name = request.get_json()["name"]
+    account_name = ""
+    if "name" not in request.get_json() and not "username" in request.get_json():
+        abort(400, "Missing name or username")
+    elif "name" not in request.get_json():
+        account_name = request.get_json()["username"]
+    else:
+        account_name = request.get_json()["name"]
+
     account_password = request.get_json()["password"]
     user = Users.query.filter_by(name=account_name).first()
     token = f"Bearer {auth_token}"
     actual_time_in_seconds = int(time.time())
     all_auth_tokens[token] = {"user": account_name, "time": actual_time_in_seconds}
     if user:
+        user_object = user.__dict__
+        del user_object["_sa_instance_state"]
+
         if user.account_type == "Kid":
             generate_log(request, "LOGIN")
+            execute_event(Events.LOGIN, user)
             return jsonify(
-                {"id": user.id, "name": user.name, "error": "None", "token": auth_token}
+                {"id": user.id, "name": user.name, "error": "None", "token": auth_token, "user": user_object}
             )
         elif user.verify_password(account_password):
             generate_log(request, "LOGIN")
+            execute_event(Events.LOGIN, user)
             return jsonify(
-                {"id": user.id, "name": user.name, "error": "None", "token": auth_token}
+                {"id": user.id, "name": user.name, "error": "None", "token": auth_token, "user": user_object}
             )
         else:
             generate_log(request, "ERROR")
+            user = user.__dict__
+            user["error"] = "Unauthorized"
+            execute_event(Events.LOGIN, user)
             return jsonify({"error": "Unauthorized"})
     else:
         generate_log(request, "ERROR")
+        user = user.__dict__
+        user["error"] = "Unauthorized"
+        execute_event(Events.LOGIN, user)
         return jsonify({"error": "Unauthorized"})
 
 
@@ -72,15 +99,15 @@ def create_account():
     account_password = body["password"]
     account_type_input = body["type"]
 
-    profil_picture = f"{IMAGES_PATH}/avatars/{account_name}.webp"
-    if "profil_picture" not in body:
-        profil_picture = "/static/img/avatars/defaultUserProfilePic.png"
+    profile_picture = f"{IMAGES_PATH}/avatars/{account_name}.webp"
+    if "profile_picture" not in body:
+        profile_picture = "/static/img/avatars/defaultUserProfilePic.png"
     else:
-        file_base64 = body["profil_picture"]
+        file_base64 = body["profile_picture"]
         if file_base64.startswith("data:image"):
             file_base64 = file_base64.split(",", 1)[1]
 
-        full_path = profil_picture
+        full_path = profile_picture
 
         image_data = base64.b64decode(file_base64)
 
@@ -114,7 +141,7 @@ def create_account():
     new_user = Users(
         name=account_name,
         password=account_password,
-        profil_picture=profil_picture,
+        profile_picture=profile_picture,
         account_type=account_type_input,
     )
     DB.session.add(new_user)
@@ -128,7 +155,7 @@ def create_account():
 
 
 @users_bp.route("/edit_profil", methods=["POST"])
-def edit_profil():
+def edit_profil() -> Response:
     authorization = request.headers.get("Authorization")
 
     if authorization not in all_auth_tokens:
@@ -158,12 +185,12 @@ def edit_profil():
     user = Users.query.filter_by(name=username_in_tokens).first()
     try:
         f = request.files["image"]
-        name, extension = os.path.splitext(f.filename)
-        profil_picture = f"/static/img/{user_name}{extension}"
+        name, extension = os.path.splitext(f.filename or "")
+        profile_picture = f"/static/img/{user_name}{extension}"
         if extension == "":
-            profil_picture = "/static/img/avatars/defaultUserProfilePic.png"
+            profile_picture = "/static/img/avatars/defaultUserProfilePic.png"
     except Exception:
-        profil_picture = "/static/img/avatars/defaultUserProfilePic.png"
+        profile_picture = "/static/img/avatars/defaultUserProfilePic.png"
 
     user_to_edit = Users.query.filter_by(id=id).first()
 
@@ -174,17 +201,17 @@ def edit_profil():
         user_to_edit.account_type = type
 
     if user_to_edit.password != generate_password_hash(password) and len(password) > 0:
-            user_to_edit.password = generate_password_hash(password)
-    
+        user_to_edit.password = generate_password_hash(password)
+
     if password == "":
         user_to_edit.password = None
     if (
-        user_to_edit.profil_picture != profil_picture
-        and "/static/img/avatars/defaultUserProfilePic.png" not in profil_picture
+        user_to_edit.profile_picture != profile_picture
+        and "/static/img/avatars/defaultUserProfilePic.png" not in profile_picture
     ):
-        f = request.files["profil_picture"]
-        f.save(f"{dir_path}{profil_picture}")
-        user_to_edit.profil_picture = profil_picture
+        f = request.files["profile_picture"]
+        f.save(f"{dir_path}{profile_picture}")
+        user_to_edit.profile_picture = profile_picture
 
     DB.session.commit()
 
@@ -197,17 +224,21 @@ def edit_profil():
 
 
 @users_bp.route("/delete_account", methods=["POST"])
-def delete_account():
+def delete_account() -> Response:
     authorization = request.headers.get("Authorization")
     check_authorization(request, authorization)
-    print(authorization)
     body = request.get_json()
+    if not body:
+        abort(400, "Missing body")
+    if "id" not in body:
+        abort(400, "Missing id")
     id = body["id"]
-    print(id)
 
     user = Users.query.filter_by(id=id).first()
     DB.session.delete(user)
     DB.session.commit()
+
+    execute_event(Events.USER_DELETE, user)
 
     return jsonify(
         {
@@ -218,21 +249,21 @@ def delete_account():
 
 
 @users_bp.route("/get_profil/<id>")
-def get_profil(id):
+def get_profil(id: int) -> Response:
     user = Users.query.filter_by(id=id).first()
-    profil_picture = user.profil_picture
-    if not os.path.exists(profil_picture):
-        profil_picture = "/static/img/avatars/defaultUserProfilePic.png"
+    profile_picture = user.profile_picture
+    if not os.path.exists(profile_picture):
+        profile_picture = "/static/img/avatars/defaultUserProfilePic.png"
     user_dict = {
         "name": user.name,
-        "profil_picture": profil_picture,
+        "profile_picture": profile_picture,
         "account_type": user.account_type,
     }
     return jsonify(user_dict)
 
 
 @users_bp.route("/is_admin", methods=["GET"])
-def is_admin():
+def is_admin() -> Response:
     authorization = request.headers.get("Authorization")
     check_authorization(request, authorization)
     user = Users.query.filter_by(name=all_auth_tokens[authorization]["user"]).first()
@@ -243,13 +274,13 @@ def is_admin():
 
 
 @users_bp.route("/invite_exist/<hash>", methods=["GET"])
-def invite_exist(hash):
+def invite_exist(hash: str) -> Response:
     can = InviteCodes.query.filter_by(code=hash).first() is not None
     return jsonify(can)
 
 
 @users_bp.route("/create_invite", methods=["POST"])
-def create_invite():
+def create_invite() -> Response:
     authorization = request.headers.get("Authorization")
     check_authorization(request, authorization)
     user = Users.query.filter_by(name=all_auth_tokens[authorization]["user"]).first()
